# app/server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CharacterRelRequest:
    properties:
      bookId:
        type: integer
      characterAId:
        type: integer
      characterBId:
        type: integer
      relationDesc:
        type: string
      relationType:
        maxLength: 32
        type: string
      sortOrder:
        type: integer
    required:
    - bookId
    - characterAId
    - characterBId
    - relationType
    type: object
  dto.CharacterRelResponse:
    properties:
      bookId:
        type: integer
      characterAId:
        type: integer
      characterAName:
        type: string
      characterBId:
        type: integer
      characterBName:
        type: string
      createBy:
        type: integer
      createTime:
        type: string
      id:
        type: integer
      relationDesc:
        type: string
      relationType:
        type: string
      sortOrder:
        type: integer
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  dto.CharacterRequest:
    properties:
      alias:
        type: string
      avatar:
        type: string
      bookId:
        type: integer
      extra:
        type: string
      intro:
        type: string
      name:
        maxLength: 128
        type: string
      roleType:
        enum:
        - "1"
        - "2"
        - "3"
        - "4"
        type: string
      sortOrder:
        type: integer
    required:
    - bookId
    - name
    - roleType
    type: object
  dto.CharacterResponse:
    properties:
      alias:
        type: string
      avatar:
        type: string
      bookId:
        type: integer
      createBy:
        type: integer
      createTime:
        type: string
      extra:
        type: string
      id:
        type: integer
      intro:
        type: string
      name:
        type: string
      roleType:
        type: string
      sortOrder:
        type: integer
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  dto.CharacterSearch:
    properties:
      bookId:
        type: integer
      current:
        type: integer
      keyword:
        type: string
      roleType:
        type: string
      size:
        type: integer
    required:
    - bookId
    type: object
...
  dto.PageResponse:
    properties:
      current:
        type: integer
      records: {}
      size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 小说阅读平台后端 API
  title: Boread API
  version: "1.0"
paths:
...
  /api/book/character:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CharacterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CharacterResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - book-character
  /api/book/character/{id}:
    delete:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - book-character
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CharacterResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - book-character
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CharacterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CharacterResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新角色
      tags:
      - book-character
  /api/book/character/book/{bookId}:
    get:
      parameters:
      - description: 书籍ID
        in: path
        name: bookId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CharacterResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 按书查询角色列表
      tags:
      - book-character
  /api/book/character/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 搜索参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CharacterSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PageResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询角色
      tags:
      - book-character
  /api/book/character/rel:
    post:
      consumes:
      - application/json
      parameters:
      - description: 关系参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CharacterRelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CharacterRelResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建角色关系
      tags:
      - book-character-rel
  /api/book/character/rel/{id}:
    delete:
      parameters:
      - description: 关系ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除角色关系
      tags:
      - book-character-rel
    get:
      parameters:
      - description: 关系ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CharacterRelResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取角色关系详情
      tags:
      - book-character-rel
  /api/book/character/rel/book/{bookId}:
    get:
      parameters:
      - description: 书籍ID
        in: path
        name: bookId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CharacterRelResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 查询某本书所有角色关系
      tags:
      - book-character-rel
  /api/book/character/rel/character/{characterId}:
    get:
      parameters:
      - description: 角色ID
        in: path
        name: characterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CharacterRelResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 查询角色所有关系
      tags:
      - book-character-rel
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"